Basic acceptance tests now pass with the behavior
