+++ ripple.config
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<ripple xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+<ripple xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <Name>NServiceBus.NHibernate</Name>
   <NugetSpecFolder>packaging/nuget</NugetSpecFolder>
   <SourceFolder>src</SourceFolder>
